homework_1
添加了一些注释
